Version 2.2.0 * An expired reminder, still sends a reminder message when available to (with an apology) * All punishments are now logged * The way of handling temp-mutes has been re-written to use the Mongo database and not a dependancy * Moderation commands now base off permissions, not the staff role * Made event utils a global variable * Fixed issue that caused a crash when doing clearuser command * Fixed a crash that occured when the userjoin role didn't exist anymore * Bot now doesn't crash when a user is punished and their DMs are closed * Reminder command will now send a test DM to check DMs are open * Trycatch statements have been added to prevent the bot from crashing on closed DMs for a reminder * Join and leave guild embeds are now fixed * Reminders can now succesfully be removed

+++ commands/mod/clear.js
@@ -8,55 +8,49 @@ module.exports = {
   run: async (bot, message, args) => {
 
     const Discord = require("discord.js");
+    const config = await bot.mutils.getGuildById(message.guild.id);
 
-    var config = await bot.mutils.getGuildById(message.guild.id);
-
+    //Perm Check
     if (!message.member.permissions.has("MANAGE_MESSAGES")) {
       return bot.noPermsEmbed(`${message.guild.name}`, bot);
     }
 
-    var amount = args[0];
+    //Args check
+    let amount = args[0];
     if (!amount || amount == undefined || isNaN(amount) || args[0] == "help") {
       return bot.helpEmbed("clear", bot)
         .then((embed) => message.reply(embed))
         .catch((error) => bot.log.post("error", error));
     }
 
-    if (amount > 100) {
-      return bot.createEmbed("error", "", `Error! You can't clear more than 100 messages at a time!`, [], `${message.guild.name}`, message)
+    if (amount > 100 || amount < 1) {
+      return bot.createEmbed("error", "", `The amount of messages to clear must be in-between 1 and 100.`, [], `${message.guild.name}`, message)
         .then((embed) => message.reply(embed))
         .catch((error) => bot.log.post("error", error));
     }
 
-    if (amount < 1) {
-      return bot.createEmbed("error", "", `Error! You can't clear less than 1 message you silly goose!`, [], `${message.guild.name}`, message)
-        .then((embed) => message.reply(embed))
-        .catch((error) => bot.log.post("error", error));
-    }
-
+    //Fetch the messages
     message.channel.messages.fetch({
       limit: amount,
     }).then((messages) => {
+      //Bulk delete them
       message.delete();
       message.channel.bulkDelete(messages).catch(error => bot.log.post("error", error.stack));
     });
 
-    message.channel.send({
-      embeds: [{
-        color: bot.settings.color.green,
-        description: `Successfully cleared **${amount}** messages in **${message.channel.name}**`
-      }]
-    });
+    //Send success message
+    bot.createEmbed("success", "", `Successfully cleared **${amount}** messages.`, [], `${message.guild.name}`, message)
+          .then((embed) => message.reply(embed))
+          .catch((error) => bot.log.post("error", error))
 
     //Logging
-    const efunctions = require('../../main/functions/eventUtils.js');
     if (config.logging.enabled == true) {
       if (config.logging.level == "low" || config.logging.level == "medium" || config.logging.level == "high") {
-        if (efunctions.checkChannel(config.logging.channel, bot) == true) {
+        if (bot.efunctions.checkChannel(config.logging.channel, bot) == true) {
           let lchannel = bot.channels.cache.get(config.logging.channel);
           bot.eventEmbed("c70011", message.author, "Bulk Delete", `**Amount:** ${amount}\n**Channel:** ${message.channel.name}`, [], `${message.guild.name}`, bot)
             .then(embed => lchannel.send(embed))
-            .catch(error => console.error(error))
+            .catch(error => bot.log.post("error", error))
         }
       }
     }
